Pedantic pedantic pedantic.

+++ cplof/src/packrat.c
@@ -317,7 +317,7 @@ struct ParseResult **packratNonterminal(struct ParseContext *ctx,
 
     if (packratWarnAmbiguous && result.bufused > 2) {
         fprintf(stderr, "Ambiguity: Production %s returned %d at %s:%d:%d\n",
-                (char *) production->name, result.bufused - 1,
+                (char *) production->name, (int) result.bufused - 1,
                 (char *) file, (int) line + 1, (int) col + 1);
     }
